Enhance/Improve the Program logging and tracing infrastructure. #590
- refactoring SemanticData.UANodeSetValidation - refactoring UT
- ready for testing
