+++ SemanticData/Tests/USNodeSetValidationUnitTestProject/AddressSpaceContextUnitTest.cs
@@ -1,15 +1,14 @@
-﻿//___________________________________________________________________________________
+﻿//__________________________________________________________________________________________________
 //
 //  Copyright (C) 2021, Mariusz Postol LODZ POLAND.
 //
-//  To be in touch join the community at GITTER: https://gitter.im/mpostol/OPC-UA-OOI
-//___________________________________________________________________________________
+//  To be in touch join the community at GitHub: https://github.com/mpostol/OPC-UA-OOI/discussions
+//__________________________________________________________________________________________________
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using UAOOI.SemanticData.BuildingErrorsHandling;
@@ -30,7 +29,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     public void AddressSpaceContextConstructorTest()
     {
       List<IUANodeBase> _invalidNodes = new List<IUANodeBase>();
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       _asp.AddressSpaceContext.UTAddressSpaceCheckConsistency(x => _invalidNodes.Add(x));
       _asp.TestConsistency(4, 0);
     }
@@ -39,7 +38,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     [TestCategory("AddressSpaceContext")]
     public void ReferencesCheckConsistencyTest()
     {
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       _asp.AddressSpaceContext.UTReferencesCheckConsistency((x, y, z, v) => Assert.Fail());
       _asp.TestConsistency(4, 0);
     }
@@ -48,7 +47,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     [TestCategory("AddressSpaceContext")]
     public void AddressSpaceContextContentCheck()
     {
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       List<IUANodeContext> _content = new List<IUANodeContext>();
       _asp.AddressSpaceContext.UTTryGetUANodeContext(VariableTypes.PropertyType, x => _content.Add(x));
       Assert.AreEqual<int>(1, _content.Count);
@@ -71,7 +70,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     [TestCategory("AddressSpaceContext")]
     public void AddressSpaceReferencesContentCheck()
     {
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       List<UAReferenceContext> _content = new List<UAReferenceContext>();
       _asp.AddressSpaceContext.UTGetReferences(ObjectIds.RootFolder, x => _content.Add(x));
       Assert.AreEqual<int>(4, _content.Count);
@@ -126,28 +125,26 @@ namespace UAOOI.SemanticData.UANodeSetValidation
           }
          }
       };
-      using (TracedAddressSpaceContext tracedAddressSpace = new TracedAddressSpaceContext())
-      {
-        AddressSpaceContext asp = new AddressSpaceContext(tracedAddressSpace);
-        ((IAddressSpaceContext)asp).ImportUANodeSet(newNodeSet);
-        List<UAReferenceContext> references = new List<UAReferenceContext>();
-        asp.UTGetReferences(NodeId.Parse(newNodeSet.Items[0].NodeId), x => references.Add(x));
-        Assert.AreEqual<int>(1, references.Count);
-        Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasProperty, references[0].ReferenceKind);
-        Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasProperty, references[0].ReferenceKind);
-        references.Clear();
-        asp.UTGetReferences(NodeId.Parse(newNodeSet.Items[1].NodeId), x => references.Add(x));
-        Assert.AreEqual<int>(2, references.Count);
-        Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasTypeDefinition, references[0].ReferenceKind);
-        Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasModellingRule, references[1].ReferenceKind);
-      }
+      Helpers.TracedAddressSpaceContext tracedAddressSpace = new Helpers.TracedAddressSpaceContext();
+      AddressSpaceContext asp = new AddressSpaceContext(tracedAddressSpace);
+      ((IAddressSpaceContext)asp).ImportUANodeSet(newNodeSet);
+      List<UAReferenceContext> references = new List<UAReferenceContext>();
+      asp.UTGetReferences(NodeId.Parse(newNodeSet.Items[0].NodeId), x => references.Add(x));
+      Assert.AreEqual<int>(1, references.Count);
+      Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasProperty, references[0].ReferenceKind);
+      Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasProperty, references[0].ReferenceKind);
+      references.Clear();
+      asp.UTGetReferences(NodeId.Parse(newNodeSet.Items[1].NodeId), x => references.Add(x));
+      Assert.AreEqual<int>(2, references.Count);
+      Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasTypeDefinition, references[0].ReferenceKind);
+      Assert.AreEqual<ReferenceKindEnum>(ReferenceKindEnum.HasModellingRule, references[1].ReferenceKind);
     }
 
     [TestMethod]
     [TestCategory("AddressSpaceContext")]
     public void AddressSpaceContextValidateAndExportModelOpcUa()
     {
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       ((IAddressSpaceContext)_asp.AddressSpaceContext).ValidateAndExportModel(new Uri(UAInformationModel.Namespaces.OpcUa));
       _asp.TestConsistency(683, 2);
       Assert.AreEqual<string>(BuildError.WrongReference2Property.Identifier, _asp.TraceList[0].BuildError.Identifier);
@@ -159,7 +156,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     [ExpectedException(typeof(ArgumentOutOfRangeException))]
     public void AddressSpaceContextValidateAndExportModelWrongNamespace()
     {
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       ((IAddressSpaceContext)_asp.AddressSpaceContext).ValidateAndExportModel(new Uri("http://www.example.com/afterthought/box"));
     }
 
@@ -167,7 +164,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     [TestCategory("AddressSpaceContext")]
     public void AddressSpaceContextValidateAndExportIndex0()
     {
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       IEnumerable<IUANodeContext> _returnValue = null;
       _asp.AddressSpaceContext.UTValidateAndExportModel(0, x => _returnValue = x);
       Assert.AreEqual<int>(3909, (_returnValue.Count<IUANodeContext>()));
@@ -181,7 +178,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     [TestCategory("AddressSpaceContext")]
     public void ImportObjectTest()
     {
-      AddressSpaceWrapper _asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext _asp = new TracedAddressSpaceContext();
       UANodeSet _newNodeSet = TestData.CreateNodeSetModel();
       ((IAddressSpaceContext)_asp.AddressSpaceContext).ImportUANodeSet(_newNodeSet);
       _asp.TestConsistency(7, 1);
@@ -228,7 +225,7 @@ namespace UAOOI.SemanticData.UANodeSetValidation
           }
          }
       };
-      AddressSpaceWrapper asp = new AddressSpaceWrapper();
+      TracedAddressSpaceContext asp = new TracedAddressSpaceContext();
       ((IAddressSpaceContext)asp.AddressSpaceContext).ImportUANodeSet(newNodeSet);
       IUANodeContext uaObjectType = asp.AddressSpaceContext.GetOrCreateNodeContext(NodeId.Parse(newNodeSet.Items[0].NodeId), x => { Assert.Fail(); return null; });
       Assert.IsNotNull(uaObjectType);
@@ -244,54 +241,14 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     [TestMethod]
     public void GetBaseTypesTest()
     {
-      using (TracedAddressSpaceContext tasp = new TracedAddressSpaceContext())
-      {
-        AddressSpaceContext asp = (AddressSpaceContext)tasp.CreateAddressSpaceContext();
-        List<IUANodeContext> inheritanceChain = new List<IUANodeContext>();
-        IUANodeContext hasPropertyNode = asp.GetOrCreateNodeContext(ReferenceTypeIds.HasProperty, x => { Assert.Fail(); return null; });
-        asp.GetBaseTypes(hasPropertyNode, inheritanceChain);
-        Assert.AreEqual<int>(5, inheritanceChain.Count);
-        Assert.AreEqual<string>(ReferenceTypeIds.HasProperty.ToString(), inheritanceChain[0].NodeIdContext.ToString());
-        Assert.AreEqual<string>(ReferenceTypeIds.References.ToString(), inheritanceChain[4].NodeIdContext.ToString());
-      }
+      Helpers.TracedAddressSpaceContext tasp = new Helpers.TracedAddressSpaceContext();
+      AddressSpaceContext asp = (AddressSpaceContext)tasp.AddressSpaceContext;
+      List<IUANodeContext> inheritanceChain = new List<IUANodeContext>();
+      IUANodeContext hasPropertyNode = asp.GetOrCreateNodeContext(ReferenceTypeIds.HasProperty, x => { Assert.Fail(); return null; });
+      asp.GetBaseTypes(hasPropertyNode, inheritanceChain);
+      Assert.AreEqual<int>(5, inheritanceChain.Count);
+      Assert.AreEqual<string>(ReferenceTypeIds.HasProperty.ToString(), inheritanceChain[0].NodeIdContext.ToString());
+      Assert.AreEqual<string>(ReferenceTypeIds.References.ToString(), inheritanceChain[4].NodeIdContext.ToString());
     }
-
-    #region private
-
-    private class AddressSpaceWrapper : IBuildErrorsHandling
-    {
-      public AddressSpaceWrapper()
-      {
-        AddressSpaceContext = new AddressSpaceContext(this);
-      }
-
-      public void TestConsistency(int diagnosticCounter, int errorsCounter)
-      {
-        Assert.AreEqual<int>(diagnosticCounter, Errors);
-        Assert.AreEqual<int>(errorsCounter, TraceList.Count);
-      }
-
-      internal List<TraceMessage> TraceList = new List<TraceMessage>();
-      internal AddressSpaceContext AddressSpaceContext = null;
-      public int Errors { get; set; } = 0;
-
-      public void WriteTraceMessage(TraceMessage traceMessage)
-      {
-        Console.WriteLine(traceMessage.ToString());
-        Errors++;
-        if (traceMessage.BuildError.Focus != Focus.Diagnostic)
-          TraceList.Add(traceMessage);
-      }
-
-      public void TraceData(TraceEventType eventType, int id, object data)
-      {
-        string message = $"TraceData eventType = {eventType}, id = {id}, {data}";
-        Console.WriteLine(message);
-        if (eventType == TraceEventType.Critical || eventType == TraceEventType.Error)
-          throw new ApplicationException(message);
-      }
-    }
-
-    #endregion private
   }
 }
