+++ SemanticData/Tests/USNodeSetValidationUnitTestProject/XMLModelsModelsWithErrorsUnitTest.cs
@@ -1,9 +1,9 @@
-﻿//___________________________________________________________________________________
+﻿//__________________________________________________________________________________________________
 //
 //  Copyright (C) 2021, Mariusz Postol LODZ POLAND.
 //
-//  To be in touch join the community at GITTER: https://gitter.im/mpostol/OPC-UA-OOI
-//___________________________________________________________________________________
+//  To be in touch join the community at GitHub: https://github.com/mpostol/OPC-UA-OOI/discussions
+//__________________________________________________________________________________________________
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
@@ -26,18 +26,16 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongEventNotifier.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(2, _log.TraceList.Count);
-        Assert.AreEqual<string>(BuildError.WrongEventNotifier.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(2, _log.TraceList.Count);
+      Assert.AreEqual<string>(BuildError.WrongEventNotifier.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -46,20 +44,18 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongReference2Property.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Count());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(3, _log.TraceList.Count());
-        Assert.AreEqual<int>(1, (_log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P0-0001050000").Count<TraceMessage>()));
-        Assert.AreEqual<string>(BuildError.WrongReference2Variable.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.WrongReference2Property.Identifier, _log.TraceList[1].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[2].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Count());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(3, _log.TraceList.Count());
+      Assert.AreEqual<int>(1, (_log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P0-0001050000").Count<TraceMessage>()));
+      Assert.AreEqual<string>(BuildError.WrongReference2Variable.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.WrongReference2Property.Identifier, _log.TraceList[1].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[2].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -68,22 +64,20 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongValueRank.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(3, _log.TraceList.Count());
-        Assert.AreEqual<string>(BuildError.WrongValueRank.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.WrongValueRank.Identifier, _log.TraceList[1].BuildError.Identifier);
-        Assert.AreEqual<string>(_log.TraceList[0].Message, _log.TraceList[1].Message); //Duplicated log entry
-        Assert.AreEqual<string>("The value -4 is not supported", _log.TraceList[0].Message);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[2].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(3, _log.TraceList.Count());
+      Assert.AreEqual<string>(BuildError.WrongValueRank.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.WrongValueRank.Identifier, _log.TraceList[1].BuildError.Identifier);
+      Assert.AreEqual<string>(_log.TraceList[0].Message, _log.TraceList[1].Message); //Duplicated log entry
+      Assert.AreEqual<string>("The value -4 is not supported", _log.TraceList[0].Message);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[2].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -92,18 +86,16 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongAccessLevel.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Count());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.IsTrue(_log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0506040000").Any<TraceMessage>());
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Count());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.IsTrue(_log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0506040000").Any<TraceMessage>());
     }
 
     [TestMethod]
@@ -112,18 +104,16 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongInverseName.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(4, _log.TraceList.Count());
-        Assert.AreEqual(3, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0503020000").Count<TraceMessage>());
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(4, _log.TraceList.Count());
+      Assert.AreEqual(3, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0503020000").Count<TraceMessage>());
     }
 
     [TestMethod]
@@ -132,18 +122,16 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\DuplicatedNodeId.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(2, _log.TraceList.Count());
-        Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.NodeIdDuplicated.Identifier, _log.TraceList[1].BuildError.Identifier);
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(1, _log.TraceList.Count);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[0].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(2, _log.TraceList.Count());
+      Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.NodeIdDuplicated.Identifier, _log.TraceList[1].BuildError.Identifier);
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(1, _log.TraceList.Count);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[0].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -152,18 +140,16 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongDisplayNameLength.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.IsTrue(_log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0502050000").Any<TraceMessage>());
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.IsTrue(_log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0502050000").Any<TraceMessage>());
     }
 
     [TestMethod]
@@ -172,18 +158,16 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongWriteMask.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(3, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.WrongWriteMaskValue.Identifier).Count<TraceMessage>());
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(3, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.WrongWriteMaskValue.Identifier).Count<TraceMessage>());
     }
 
     [TestMethod]
@@ -192,19 +176,17 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\NotSupportedFeature.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        Assert.IsNotNull(_as);
-        Assert.IsNotNull(_as);
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[0].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      Assert.IsNotNull(_as);
+      Assert.IsNotNull(_as);
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[0].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -213,19 +195,17 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongBrowseName.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(2, _log.TraceList.Count());
-        Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.EmptyBrowseName.Identifier, _log.TraceList[1].BuildError.Identifier);
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.AreEqual<string>(BuildError.WrongSymbolicName.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(2, _log.TraceList.Count());
+      Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.EmptyBrowseName.Identifier, _log.TraceList[1].BuildError.Identifier);
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(2, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.AreEqual<string>(BuildError.WrongSymbolicName.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -234,17 +214,15 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\WrongNodeId.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(2, _log.TraceList.Count());
-        Assert.AreEqual<string>(BuildError.NodeIdNotDefined.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(2, _log.TraceList.Count());
+      Assert.AreEqual<string>(BuildError.NodeIdNotDefined.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -253,17 +231,15 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\UndefinedHasSubtype.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(2, _log.TraceList.Count());
-        Assert.AreEqual<string>(BuildError.UndefinedHasSubtypeTarget.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(2, _log.TraceList.Count());
+      Assert.AreEqual<string>(BuildError.UndefinedHasSubtypeTarget.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, _log.TraceList[1].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -272,17 +248,15 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\UndefinedHasTypeDefinition.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        _as.ValidateAndExportModel(m_NameSpace);
-        //TODO Recognize problems with P3.7.13 HasTypeDefinition ReferenceType #39
-        Assert.Inconclusive("Instances are not imported - the error is not recognized.");
-        Assert.AreEqual<int>(3, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0713000000").Count<TraceMessage>());
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      _as.ValidateAndExportModel(m_NameSpace);
+      //TODO Recognize problems with P3.7.13 HasTypeDefinition ReferenceType #39
+      Assert.Inconclusive("Instances are not imported - the error is not recognized.");
+      Assert.AreEqual<int>(3, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.AreEqual<int>(1, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == "P3-0713000000").Count<TraceMessage>());
     }
 
     /// <summary>
@@ -294,19 +268,17 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ModelsWithErrors\UndefinedHasChildren.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext _log = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext _as = _log.CreateAddressSpaceContext();
-        _as.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, _log.TraceList.Count());
-        _log.Clear();
-        _as.ValidateAndExportModel(m_NameSpace);
-        Assert.AreEqual<int>(5, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
-        Assert.AreEqual<string>(BuildError.NodeCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.NodeCannotBeNull.Identifier, _log.TraceList[1].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.DanglingReferenceTarget.Identifier, _log.TraceList[2].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.DanglingReferenceTarget.Identifier, _log.TraceList[3].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext _log = new TracedAddressSpaceContext();
+      IAddressSpaceContext _as = _log.AddressSpaceContext;
+      _as.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, _log.TraceList.Count());
+      _log.Clear();
+      _as.ValidateAndExportModel(m_NameSpace);
+      Assert.AreEqual<int>(5, _log.TraceList.Where<TraceMessage>(x => x.BuildError.Focus != Focus.Diagnostic).Count<TraceMessage>());
+      Assert.AreEqual<string>(BuildError.NodeCannotBeNull.Identifier, _log.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.NodeCannotBeNull.Identifier, _log.TraceList[1].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.DanglingReferenceTarget.Identifier, _log.TraceList[2].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.DanglingReferenceTarget.Identifier, _log.TraceList[3].BuildError.Identifier);
     }
 
     #endregion TestMethod
