+++ SemanticData/Tests/USNodeSetValidationUnitTestProject/XMLModelsProblemsToReportUnitTest.cs
@@ -26,29 +26,27 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ProblemsToReport\ADI#509\Opc.Ua.Adi.NodeSet2.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext addressSpace = traceContext.CreateAddressSpaceContext();
-        addressSpace.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(1, traceContext.TraceList.Count);
-        Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, traceContext.TraceList[0].BuildError.Identifier);
-        traceContext.Clear();
-        addressSpace.ValidateAndExportModel(new UriBuilder("http://opcfoundation.org/UA/ADI/").Uri);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.DataEncoding).Count<TraceMessage>());
-        Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.DataType).Count<TraceMessage>());
-        Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.Naming).Count<TraceMessage>());
-        Assert.AreEqual<int>(24, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.NodeClass).Count<TraceMessage>());
-        Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.NonCategorized).Count<TraceMessage>());
-        Assert.AreEqual<int>(23, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.Reference).Count<TraceMessage>());
-        Assert.AreEqual<int>(2, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.XML).Count<TraceMessage>());
-        //errors
-        Assert.AreEqual<int>(49, traceContext.TraceList.Count);
-        Assert.AreEqual<int>(3, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.NodeIdNotDefined.Identifier).Count<TraceMessage>());
-        Assert.AreEqual<int>(5, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.UndefinedHasSubtypeTarget.Identifier).Count<TraceMessage>());
-        Assert.AreEqual<int>(18, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.UndefinedHasTypeDefinition.Identifier).Count<TraceMessage>());
-        Assert.AreEqual<int>(1, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.ModelContainsErrors.Identifier).Count<TraceMessage>());
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, traceContext.TraceList[48].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext();
+      IAddressSpaceContext addressSpace = traceContext.AddressSpaceContext;
+      addressSpace.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(1, traceContext.TraceList.Count);
+      Assert.AreEqual<string>(BuildError.ModelsCannotBeNull.Identifier, traceContext.TraceList[0].BuildError.Identifier);
+      traceContext.Clear();
+      addressSpace.ValidateAndExportModel(new UriBuilder("http://opcfoundation.org/UA/ADI/").Uri);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.DataEncoding).Count<TraceMessage>());
+      Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.DataType).Count<TraceMessage>());
+      Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.Naming).Count<TraceMessage>());
+      Assert.AreEqual<int>(24, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.NodeClass).Count<TraceMessage>());
+      Assert.AreEqual<int>(0, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.NonCategorized).Count<TraceMessage>());
+      Assert.AreEqual<int>(23, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.Reference).Count<TraceMessage>());
+      Assert.AreEqual<int>(2, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Focus == Focus.XML).Count<TraceMessage>());
+      //errors
+      Assert.AreEqual<int>(49, traceContext.TraceList.Count);
+      Assert.AreEqual<int>(3, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.NodeIdNotDefined.Identifier).Count<TraceMessage>());
+      Assert.AreEqual<int>(5, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.UndefinedHasSubtypeTarget.Identifier).Count<TraceMessage>());
+      Assert.AreEqual<int>(18, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.UndefinedHasTypeDefinition.Identifier).Count<TraceMessage>());
+      Assert.AreEqual<int>(1, traceContext.TraceList.Where<TraceMessage>(x => x.BuildError.Identifier == BuildError.ModelContainsErrors.Identifier).Count<TraceMessage>());
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, traceContext.TraceList[48].BuildError.Identifier);
     }
 
     [TestMethod]
@@ -56,40 +54,38 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ProblemsToReport\BrowseNameInheritedFrom0\BrowseNameInheritedFrom0.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext())
+      TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext();
+      IAddressSpaceContext addressSpace = traceContext.AddressSpaceContext;
+      ModelFactoryTestingFixture.InformationModelFactoryBase testingModelFixture = new InformationModelFactoryBase();
+      addressSpace.InformationModelFactory = testingModelFixture;
+      addressSpace.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Count);
+      traceContext.Clear();
+      addressSpace.ValidateAndExportModel(new UriBuilder("http://tricycleTypeV1").Uri);
+      Assert.AreEqual<int>(2, traceContext.TraceList.Count);
+      IEnumerable<NodeFactoryBase> nodes = testingModelFixture.Export();
+      Assert.AreEqual(3, nodes.Count<NodeFactoryBase>());
+      Dictionary<string, NodeFactoryBase> nodesDictionary = nodes.ToDictionary<NodeFactoryBase, string>(x => x.SymbolicName.Name);
+      AddressSpaceContext asContext = addressSpace as AddressSpaceContext;
+      IEnumerable<IUANodeContext> allNodes = null;
+      asContext.UTValidateAndExportModel(1, x => allNodes = x);
+      Assert.IsNotNull(allNodes);
+      List<IUANodeContext> orphanedNodes = new List<IUANodeContext>();
+      List<IUANodeContext> processedNodes = new List<IUANodeContext>();
+      foreach (IUANodeContext item in allNodes)
       {
-        IAddressSpaceContext addressSpace = traceContext.CreateAddressSpaceContext();
-        ModelFactoryTestingFixture.InformationModelFactoryBase testingModelFixture = new InformationModelFactoryBase();
-        addressSpace.InformationModelFactory = testingModelFixture;
-        addressSpace.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Count);
-        traceContext.Clear();
-        addressSpace.ValidateAndExportModel(new UriBuilder("http://tricycleTypeV1").Uri);
-        Assert.AreEqual<int>(2, traceContext.TraceList.Count);
-        IEnumerable<NodeFactoryBase> nodes = testingModelFixture.Export();
-        Assert.AreEqual(3, nodes.Count<NodeFactoryBase>());
-        Dictionary<string, NodeFactoryBase> nodesDictionary = nodes.ToDictionary<NodeFactoryBase, string>(x => x.SymbolicName.Name);
-        AddressSpaceContext asContext = addressSpace as AddressSpaceContext;
-        IEnumerable<IUANodeContext> allNodes = null;
-        asContext.UTValidateAndExportModel(1, x => allNodes = x);
-        Assert.IsNotNull(allNodes);
-        List<IUANodeContext> orphanedNodes = new List<IUANodeContext>();
-        List<IUANodeContext> processedNodes = new List<IUANodeContext>();
-        foreach (IUANodeContext item in allNodes)
+        if (!nodesDictionary.ContainsKey(item.UANode.BrowseNameQualifiedName.Name))
         {
-          if (!nodesDictionary.ContainsKey(item.UANode.BrowseNameQualifiedName.Name))
-          {
-            orphanedNodes.Add(item);
-            Debug.WriteLine($"The following node has been removed from the model: {item.ToString()}");
-          }
-          else
-            processedNodes.Add(item);
+          orphanedNodes.Add(item);
+          Debug.WriteLine($"The following node has been removed from the model: {item.ToString()}");
         }
-        Debug.WriteLine($"The recovered information model contains {nodesDictionary.Count} nodes");
-        Debug.WriteLine($"The source information model contains {allNodes.Count<IUANodeContext>()} nodes");
-        Debug.WriteLine($"Number of nodes not considered for export {orphanedNodes.Count}");
-        Debug.WriteLine($"Number of processed nodes {processedNodes.Count}");
+        else
+          processedNodes.Add(item);
       }
+      Debug.WriteLine($"The recovered information model contains {nodesDictionary.Count} nodes");
+      Debug.WriteLine($"The source information model contains {allNodes.Count<IUANodeContext>()} nodes");
+      Debug.WriteLine($"Number of nodes not considered for export {orphanedNodes.Count}");
+      Debug.WriteLine($"Number of processed nodes {processedNodes.Count}");
     }
 
     [TestMethod]
@@ -97,40 +93,38 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ProblemsToReport\eoursel510\Opc.Ua.NodeSet2.TriCycleType_V1.1.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext())
+      TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext();
+      IAddressSpaceContext addressSpace = traceContext.AddressSpaceContext;
+      ModelFactoryTestingFixture.InformationModelFactoryBase testingModelFixture = new InformationModelFactoryBase();
+      addressSpace.InformationModelFactory = testingModelFixture;
+      addressSpace.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Count);
+      traceContext.Clear();
+      addressSpace.ValidateAndExportModel(new UriBuilder("http://tricycleTypeV1").Uri);
+      Assert.AreEqual<int>(5, traceContext.TraceList.Count);
+      IEnumerable<NodeFactoryBase> nodes = testingModelFixture.Export();
+      Assert.AreEqual(22, nodes.Count<NodeFactoryBase>());
+      Dictionary<string, NodeFactoryBase> nodesDictionary = nodes.ToDictionary<NodeFactoryBase, string>(x => x.SymbolicName.Name);
+      AddressSpaceContext asContext = addressSpace as AddressSpaceContext;
+      IEnumerable<IUANodeContext> allNodes = null;
+      asContext.UTValidateAndExportModel(1, x => allNodes = x);
+      Assert.IsNotNull(allNodes);
+      List<IUANodeContext> orphanedNodes = new List<IUANodeContext>();
+      List<IUANodeContext> processedNodes = new List<IUANodeContext>();
+      foreach (IUANodeContext item in allNodes)
       {
-        IAddressSpaceContext addressSpace = traceContext.CreateAddressSpaceContext();
-        ModelFactoryTestingFixture.InformationModelFactoryBase testingModelFixture = new InformationModelFactoryBase();
-        addressSpace.InformationModelFactory = testingModelFixture;
-        addressSpace.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Count);
-        traceContext.Clear();
-        addressSpace.ValidateAndExportModel(new UriBuilder("http://tricycleTypeV1").Uri);
-        Assert.AreEqual<int>(5, traceContext.TraceList.Count);
-        IEnumerable<NodeFactoryBase> nodes = testingModelFixture.Export();
-        Assert.AreEqual(22, nodes.Count<NodeFactoryBase>());
-        Dictionary<string, NodeFactoryBase> nodesDictionary = nodes.ToDictionary<NodeFactoryBase, string>(x => x.SymbolicName.Name);
-        AddressSpaceContext asContext = addressSpace as AddressSpaceContext;
-        IEnumerable<IUANodeContext> allNodes = null;
-        asContext.UTValidateAndExportModel(1, x => allNodes = x);
-        Assert.IsNotNull(allNodes);
-        List<IUANodeContext> orphanedNodes = new List<IUANodeContext>();
-        List<IUANodeContext> processedNodes = new List<IUANodeContext>();
-        foreach (IUANodeContext item in allNodes)
+        if (!nodesDictionary.ContainsKey(item.UANode.BrowseNameQualifiedName.Name))
         {
-          if (!nodesDictionary.ContainsKey(item.UANode.BrowseNameQualifiedName.Name))
-          {
-            orphanedNodes.Add(item);
-            Debug.WriteLine($"The following node has been removed from the model: {item.ToString()}");
-          }
-          else
-            processedNodes.Add(item);
+          orphanedNodes.Add(item);
+          Debug.WriteLine($"The following node has been removed from the model: {item.ToString()}");
         }
-        Debug.WriteLine($"The recovered information model contains {nodesDictionary.Count} nodes");
-        Debug.WriteLine($"The source information model contains {allNodes.Count<IUANodeContext>()} nodes");
-        Debug.WriteLine($"Number of nodes not considered for export {orphanedNodes.Count}");
-        Debug.WriteLine($"Number of processed nodes {processedNodes.Count}");
+        else
+          processedNodes.Add(item);
       }
+      Debug.WriteLine($"The recovered information model contains {nodesDictionary.Count} nodes");
+      Debug.WriteLine($"The source information model contains {allNodes.Count<IUANodeContext>()} nodes");
+      Debug.WriteLine($"Number of nodes not considered for export {orphanedNodes.Count}");
+      Debug.WriteLine($"Number of processed nodes {processedNodes.Count}");
     }
 
     [TestMethod]
@@ -138,15 +132,13 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ProblemsToReport\fgolra177\Opc.Ua.Semi.NodeSet2.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext addressSpace = traceContext.CreateAddressSpaceContext();
-        addressSpace.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Count);
-        traceContext.Clear();
-        addressSpace.ValidateAndExportModel(new UriBuilder("https://agileo-automation.com/UA/Semi/").Uri);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Count);
-      }
+      TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext();
+      IAddressSpaceContext addressSpace = traceContext.AddressSpaceContext;
+      addressSpace.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Count);
+      traceContext.Clear();
+      addressSpace.ValidateAndExportModel(new UriBuilder("https://agileo-automation.com/UA/Semi/").Uri);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Count);
     }
 
     [TestMethod]
@@ -154,15 +146,13 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ProblemsToReport\HasOrderedComponent\Opc.Ua.NodeSet2.TriCycleType_V1.1.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext addressSpace = traceContext.CreateAddressSpaceContext();
-        addressSpace.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Count);
-        traceContext.Clear();
-        addressSpace.ValidateAndExportModel(new UriBuilder("http://tricycleTypeV1").Uri);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Count);
-      }
+      TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext();
+      IAddressSpaceContext addressSpace = traceContext.AddressSpaceContext;
+      addressSpace.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Count);
+      traceContext.Clear();
+      addressSpace.ValidateAndExportModel(new UriBuilder("http://tricycleTypeV1").Uri);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Count);
     }
 
     //[TestMethod]
@@ -170,18 +160,16 @@ namespace UAOOI.SemanticData.UANodeSetValidation
     {
       FileInfo _testDataFileInfo = new FileInfo(@"ProblemsToReport\MachineVision\Opc.Ua.MachineVision.NodeSet2.xml");
       Assert.IsTrue(_testDataFileInfo.Exists);
-      using (TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext())
-      {
-        IAddressSpaceContext addressSpace = traceContext.CreateAddressSpaceContext();
-        addressSpace.ImportUANodeSet(_testDataFileInfo);
-        Assert.AreEqual<int>(0, traceContext.TraceList.Count);
-        traceContext.Clear();
-        addressSpace.ValidateAndExportModel(new UriBuilder("http://opcfoundation.org/UA/MachineVision").Uri);
-        Assert.AreEqual<int>(3, traceContext.TraceList.Count);
-        Assert.AreEqual<string>(BuildError.WrongInverseName.Identifier, traceContext.TraceList[0].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.WrongInverseName.Identifier, traceContext.TraceList[1].BuildError.Identifier);
-        Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, traceContext.TraceList[2].BuildError.Identifier);
-      }
+      TracedAddressSpaceContext traceContext = new TracedAddressSpaceContext();
+      IAddressSpaceContext addressSpace = traceContext.AddressSpaceContext;
+      addressSpace.ImportUANodeSet(_testDataFileInfo);
+      Assert.AreEqual<int>(0, traceContext.TraceList.Count);
+      traceContext.Clear();
+      addressSpace.ValidateAndExportModel(new UriBuilder("http://opcfoundation.org/UA/MachineVision").Uri);
+      Assert.AreEqual<int>(3, traceContext.TraceList.Count);
+      Assert.AreEqual<string>(BuildError.WrongInverseName.Identifier, traceContext.TraceList[0].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.WrongInverseName.Identifier, traceContext.TraceList[1].BuildError.Identifier);
+      Assert.AreEqual<string>(BuildError.ModelContainsErrors.Identifier, traceContext.TraceList[2].BuildError.Identifier);
     }
   }
 }
